GUVNOR-1208 All PageResponse objects are now built with PageResponseBuilder. Removed ServiceRowSizeHelper.

+++ guvnor-webapp/src/main/java/org/drools/guvnor/server/builder/PageResponseBuilder.java
@@ -76,7 +76,7 @@ public class PageResponseBuilder<T extends AbstractPageRow> {
         return pageResponse;
     }
 
-    public PageResponse<T> buildWithTotalRowCount(final int totalRowCount) {
+    public PageResponse<T> buildWithTotalRowCount(final long totalRowCount) {
         fixTotalRowSize( totalRowCount );
         return pageResponse;
     }
